refactor(registar): преименуј модул migration_converters → konvertori (#300)
Краће, доследно име за модул са класом `Konvertor`.

- `utils/migration_converters.py` → `utils/konvertori.py`
- `tests/test_migration_converters.py` → `tests/test_konvertori.py`
- репоинтирани увози у 3 команде/помоћника
