+++ crkva/registar/management/commands/migracija_svestenika.py
@@ -6,7 +6,7 @@ from django.db import connection
 from django.db.utils import IntegrityError
 from registar.management.commands.base_migration import MigrationCommand
 from registar.models import Parohija, Svestenik
-from registar.utils.migration_converters import Konvertor
+from registar.utils.konvertori import Konvertor
 
 
 class Command(MigrationCommand):
